fix: modify daeguk to match
