--- src/main/java/begin_a_gain/omokwang/daeguk/domain/CompletionStatus.java
@@ -1,7 +1,0 @@
-package begin_a_gain.omokwang.daeguk.domain;
-
-public enum CompletionStatus {
-    COMPLETED,
-    NOT_COMPLETED,
-    LATE_COMPLETED
-}
