--- src/main/java/begin_a_gain/omokwang/daeguk/repository/DaegukDayRepository.java
@@ -1,8 +1,0 @@
-package begin_a_gain.omokwang.daeguk.repository;
-
-import begin_a_gain.omokwang.daeguk.domain.DaegukDay;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface DaegukDayRepository extends JpaRepository<DaegukDay, Long> {
-}
