--- src/main/java/begin_a_gain/omokwang/daeguk_detail/controller/DaegukSearchController.java
@@ -1,12 +1,0 @@
-package begin_a_gain.omokwang.daeguk_detail.controller;
-
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@Tag(name = "Daeguk Search", description = "Daeguk Search API")
-@RestController
-@RequiredArgsConstructor
-public class DaegukSearchController {
-
-}
