--- src/main/java/begin_a_gain/omokwang/daeguk_detail/repository/DaegukParticipantRepository.java
@@ -1,9 +1,0 @@
-package begin_a_gain.omokwang.daeguk_detail.repository;
-
-
-import begin_a_gain.omokwang.daeguk_detail.domain.DaegukParticipant;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DaegukParticipantRepository extends JpaRepository<DaegukParticipant, Long> {
-
-}
