+++ src/main/java/begin_a_gain/omokwang/match/domain/CompletionStatus.java
@@ -1,0 +1,7 @@
+package begin_a_gain.omokwang.match.domain;
+
+public enum CompletionStatus {
+    COMPLETED,
+    NOT_COMPLETED,
+    LATE_COMPLETED
+}
