+++ src/main/java/begin_a_gain/omokwang/match/domain/MatchProgress.java
@@ -1,0 +1,36 @@
+package begin_a_gain.omokwang.match.domain;
+
+import begin_a_gain.omokwang.user.dto.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
+import java.time.LocalDate;
+
+@Entity
+@Table(name = "match_progress", uniqueConstraints = {
+        @UniqueConstraint(columnNames = {"match_id", "user_id", "progress_date"})})
+public class MatchProgress {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "match_id", nullable = false)
+    MatchInfo match;
+
+    @Column(nullable = false)
+    private LocalDate pregressDate;
+
+    private boolean completed;
+}
