+++ src/main/java/begin_a_gain/omokwang/match_detail/controller/MatchSearchController.java
@@ -1,0 +1,12 @@
+package begin_a_gain.omokwang.match_detail.controller;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@Tag(name = "Match Search", description = "Match Search API")
+@RestController
+@RequiredArgsConstructor
+public class MatchSearchController {
+
+}
